+++ Source/SmithUE/Private/Tests/SmithUECliCheckerTest.cpp
@@ -20,7 +20,7 @@
 // ---------------------------------------------------------------------------
 IMPLEMENT_SIMPLE_AUTOMATION_TEST(FTest_ParseSemver,
     "SmithUECliChecker.ParseSemver",
-    EAutomationTestFlags::ApplicationContextMask | EAutomationTestFlags::SmokeFilter)
+    EAutomationTestFlags_ApplicationContextMask | EAutomationTestFlags::SmokeFilter)
 
 bool FTest_ParseSemver::RunTest(const FString& Parameters)
 {
@@ -110,7 +110,7 @@ bool FTest_ParseSemver::RunTest(const FString& Parameters)
 // ---------------------------------------------------------------------------
 IMPLEMENT_SIMPLE_AUTOMATION_TEST(FTest_CompareSemver,
     "SmithUECliChecker.CompareSemver",
-    EAutomationTestFlags::ApplicationContextMask | EAutomationTestFlags::SmokeFilter)
+    EAutomationTestFlags_ApplicationContextMask | EAutomationTestFlags::SmokeFilter)
 
 bool FTest_CompareSemver::RunTest(const FString& Parameters)
 {
@@ -152,7 +152,7 @@ bool FTest_CompareSemver::RunTest(const FString& Parameters)
 // ---------------------------------------------------------------------------
 IMPLEMENT_SIMPLE_AUTOMATION_TEST(FTest_ClassifyVersionProbe,
     "SmithUECliChecker.ClassifyVersionProbe",
-    EAutomationTestFlags::ApplicationContextMask | EAutomationTestFlags::SmokeFilter)
+    EAutomationTestFlags_ApplicationContextMask | EAutomationTestFlags::SmokeFilter)
 
 bool FTest_ClassifyVersionProbe::RunTest(const FString& Parameters)
 {
@@ -198,7 +198,7 @@ bool FTest_ClassifyVersionProbe::RunTest(const FString& Parameters)
 // ---------------------------------------------------------------------------
 IMPLEMENT_SIMPLE_AUTOMATION_TEST(FTest_ParseNpmLsCliVersion,
     "SmithUECliChecker.ParseNpmLsCliVersion",
-    EAutomationTestFlags::ApplicationContextMask | EAutomationTestFlags::SmokeFilter)
+    EAutomationTestFlags_ApplicationContextMask | EAutomationTestFlags::SmokeFilter)
 
 bool FTest_ParseNpmLsCliVersion::RunTest(const FString& Parameters)
 {